requirements.txt
absl-py==2.4.0
anyio==4.13.0
astunparse==1.6.3
blinker==1.9.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
firebase_admin==7.4.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
flatbuffers==25.12.19
gast==0.7.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
mysql-connector-python==9.6.0
namex==0.1.0
nltk==3.9.4
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.2.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
